✨ Brave history (#41)
This plugin will parse your Brave browser history and add it to the
bottom of the centerpiece search list. This should give a nice fallback
in case no other plugin entries match your search. Furthermore, this is
another step on making the omnibox in your browser obsolete :blush:.

The list of history entries is sorted by the `number of visits to a
page` followed by the `last visit to a page`.

This resolves #24.