✨ plugins: add `$GIT_DIRECTORY_NAME` env variable
Expose the `$GIT_DIRECTORY_NAME` environment variable to the git_repositories plugin.

This is useful for setting `app_ids`, or `tmux` session names, for
example.

I am not sure about error handling here.
The unwraps are there because:
- We have a git directory, the file name should always succeed.
- It should always be valid Utf8.

We could return an empty string in an error case, that way scripts might
still succeed and fill in a default value.