⚡️ implement custom scrollable widget (#173)
Since the iced scrollable widget renders all entries passed to it, we're
experiencing degraded performance when plugins pass many entries to
render. This is noticeable when using plugins like brave or firefox
history with large history lists.

This PR removes the iced scrollable widget and instead creates a list of
10 enties which behave like a scrollable. This way we really only paint
the 10 entries which are in the current viewport.

This fixes #157.