🔎 Fuzzy matching with nucleo (#238)
Adds fuzzy matching via the
[nucleo](https://github.com/helix-editor/nucleo) crate. This should be
practically as fast, if not faster, than the previous implementation
since nucleo does a prefiltering step with `memchr`. It first fuzzy
matches against the `title` of the entry, falling back to substring
matching against the `meta`. A match on the `title` receives a `+ 1000`
to the score to always beat a match on `meta`.

Before:


![image](https://github.com/user-attachments/assets/79fa0e21-8040-4320-9302-6e663917bb05)

After:


![image](https://github.com/user-attachments/assets/27ebf626-2035-46a4-82d6-81537998753d)